Moved prod examples to howto and wrote nginx howto
