+++ reahl-dev/reahl_dev.egg-info/PKG-INFO
@@ -1,6 +1,6 @@
 Metadata-Version: 1.2
 Name: reahl-dev
-Version: 5.0.0
+Version: 5.0.1
 Summary: The core Reahl development tools.
 Home-page: http://www.reahl.org
 Maintainer: Iwan Vosloo
