+++ reahl-dev/reahl_dev.egg-info/SOURCES.txt
@@ -4,7 +4,6 @@
 COPYING
 LICENSE
 MANIFEST.in
-setup.py
 tox.ini
 reahl/__init__.py
 reahl/dev/__init__.py
