+++ reahl-doc/reahl/doc/examples/howtos/hellonginx/.reahlproject
@@ -5,6 +5,7 @@
    <egg name="reahl-web"/>
    <egg name="reahl-sqlalchemysupport"/>
    <egg name="reahl-web-declarative"/>
+   <egg name="reahl-sqlitesupport"/>
   </deps>
  </version>
  <!-- packaging -->
