+++ reahl-doc/reahl/doc/examples/howtos/hellonginx/prod/etc/mailutil.config.py
@@ -1,0 +1,1 @@
+mail.smtp_port = 25
