+++ reahl-doc/reahl/doc/examples/howtos/hellonginx/prod/etc/reahl.config.py
@@ -7,3 +7,7 @@ reahlsystem.connection_uri = 'sqlite:////var/local/hellonginx/hellonginx.db'
 
 # If using PostgreSQL:
 #reahlsystem.connection_uri = 'postgresql://hellonginx:hellonginx@localhost/hellonginx'  
+
+reahlsystem.debug = False
+
+#reahl.component.config.ConfigurationException: reahlsystem.debug has been defaulted to a value not suitable for production use: "True". You can set it in /etc/reahl.d/hellonginx/reahl.config.py
