+++ reahl-doc/reahl/doc/examples/howtos/hellonginx/prod/uwsgi/hellonginx.ini
@@ -1,9 +1,9 @@
 [uwsgi]
-plugins = python
-virtualenv = /usr/local/hellonginx/virtualenv/ 
+plugins = python3
+virtualenv = /usr/local/hellonginx/venv/
 module = hellonginxwsgi
 enable-threads = true
-processes = 2
+processes = 1
 threads = 2
 lazy = true
 harakiri = 20
