+++ reahl-doc/reahl/doc_dev/test_examples.py
@@ -28,8 +28,8 @@ from reahl.sqlalchemysupport import Session
 from reahl.webdev.tools import XPath, Browser
 
 from reahl.doc.examples.tutorial.hello.hello import HelloUI
-from reahl.doc.examples.tutorial.helloapache import helloapache
-from reahl.doc.examples.tutorial.hellonginx import hellonginx
+from reahl.doc.examples.howtos.helloapache import helloapache
+from reahl.doc.examples.howtos.hellonginx import hellonginx
 from reahl.doc.examples.tutorial.slots.slots import SlotsUI
 from reahl.doc.examples.features.tabbedpanel.tabbedpanel import TabbedPanelUI
 from reahl.doc.examples.features.carousel.carousel import CarouselUI
